Move FakeMeetup to the same package as the lsug.Meetup
