+++ server/test/src/scala/lsug/ServerSpec.scala
@@ -5,7 +5,6 @@ import lsug.{IOSuite, Server}
 
 import java.nio.file.Paths
 import java.time._
-import lsug.markup.FakeMeetup
 
 final class ServerSpec extends IOSuite {
